+++ CHANGELOG.md
@@ -90,6 +90,7 @@ This is the rolling changelog for TShock for Terraria. Use past tense when addin
 * Added filtering and validation on packet 96 (Teleport player through portal) (@QuiCM)
 * Update tracker now uses TLS (@pandabear41) 
 * When deleting an user account, any player logged in to that account is now logged out properly (@Enerdy)
+* Add NPCAddBuff data handler and bouncer (@AxeelAnder)
 
 ## TShock 4.3.25
 * Fixed a critical exploit in the Terraria protocol that could cause massive unpreventable world corruption as well as a number of other problems. Thanks to @bartico6 for reporting. Fixed by the efforts of @QuiCM, @hakusaro, and tips in the right directioon from @bartico6.